requirements.txt
asgiref==3.11.1
Django==6.0.3
django-tenants==3.10.1
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2025.3
